Working UTF8 with shuffle, scramble, etc..

+++ Server/hdrs/externs.h
@@ -49,7 +49,7 @@ typedef struct ansisplit {
 	int	i_special;	/* Special ansi characters */
 	char	c_accent;	/* Various accent characters */
         int	i_ascii8;	/* ASCII-8 encoding */
-	int i_utf8;	/* UTF8 encoding */
+	long i_utf8;	/* UTF8 encoding */
 } ANSISPLIT;
 
 typedef struct atrp {
